feat: 인기있는 학습로그 목록 스타일 정리
